setup send senz msg to pay

+++ app/src/main/java/com/score/payz/ui/PayDetailsActivity.java
@@ -82,7 +82,7 @@ public class PayDetailsActivity extends Activity implements View.OnClickListener
             //attributesList.add(new Attribute("Client NIC", transaction.getClientNic()));
             attributesList.add(new Attribute("Shop No", pay.getShopNo()));
             //attributesList.add(new Attribute("Transaction Type", transaction.getTransactionType()));
-            attributesList.add(new Attribute("Amount", Integer.toString(pay.getPayAmount())));
+            attributesList.add(new Attribute("Amount", Double.toString(pay.getPayAmount())));
             attributesList.add(new Attribute("Time", pay.getPayTime()));
 
             // populate list
